requirements.txt
flask
flask-socketio
dronekit
pyserial
pymavlink